+++ src/uniforms/mod.rs
@@ -94,9 +94,12 @@ pub use self::buffer::UniformBuffer;
 pub use self::sampler::{SamplerWrapFunction, MagnifySamplerFilter, MinifySamplerFilter};
 pub use self::sampler::{Sampler, SamplerBehavior};
 pub use self::uniforms::{EmptyUniforms, UniformsStorage};
-pub use self::value::{UniformValue, AsUniformValue, UniformType};
+pub use self::value::{UniformValue, UniformType};
 
+use buffer::Content as BufferContent;
+use buffer::BufferView;
 use program;
+use program::BlockLayout;
 
 mod bind;
 mod buffer;
@@ -105,13 +108,122 @@ mod uniforms;
 mod value;
 
 /// Object that contains the values of all the uniforms to bind to a program.
+///
+/// Objects of this type can be passed to the `draw()` function.
 pub trait Uniforms {
     /// Calls the parameter once with the name and value of each uniform.
     fn visit_values<'a, F: FnMut(&str, UniformValue<'a>)>(&'a self, F);
 }
 
-/// Objects that are suitable for being binded to a uniform block.
-pub trait UniformBlock: Copy {
-    /// Checks whether the uniforms' layout matches the given block.
-    fn matches(&program::UniformBlock) -> bool;
+/// Error about a block layout mismatch.
+#[derive(Clone, Debug)]
+pub enum LayoutMismatchError {
+    /// There is a mismatch in the type of one element.
+    TypeMismatch {
+        /// Type expected by the shader.
+        expected: UniformType,
+        /// Type that you gave.
+        obtained: UniformType,
+    },
+
+    /// The expected layout is totally different from what we have.
+    LayoutMismatch {
+        /// Layout expected by the shader.
+        expected: BlockLayout,
+        /// Layout of the input.
+        obtained: BlockLayout,
+    },
+
+    /// The type of data is good, but there is a misalignment.
+    OffsetMismatch {
+        /// Expected offset of a member.
+        expected: usize,
+        /// Offset of the same member in the input.
+        obtained: usize,
+    },
+
+    /// There is a mismatch in a submember of this layout.
+    ///
+    /// This is kind of a hierarchy inside the `LayoutMismatchError`s.
+    MemberMismatch {
+        /// Name of the field.
+        member: String,
+        /// The sub-error.
+        err: Box<LayoutMismatchError>,
+    },
+
+    /// A field is missing in either the expected of the input data layout.
+    MissingField {
+        /// Name of the field.
+        name: String,
+    },
+}
+
+/// Value that can be used as the value of a uniform.
+///
+/// This includes buffers and textures for example.
+pub trait AsUniformValue {
+    /// Builds a `UniformValue`.
+    fn as_uniform_value(&self) -> UniformValue;
+}
+
+// TODO: no way to bind a slice
+impl<'a, T: ?Sized> AsUniformValue for &'a BufferView<T> where T: UniformBlock + BufferContent {
+    fn as_uniform_value(&self) -> UniformValue {
+        fn f<T: ?Sized>(block: &program::UniformBlock)
+                        -> Result<(), LayoutMismatchError> where T: UniformBlock + BufferContent
+        {
+            // TODO: more checks?
+            T::matches(&block.layout, 0)
+        }
+
+        UniformValue::Block(self.as_slice_any(), f::<T>)
+    }
+}
+
+/// Objects that are suitable for being inside a uniform block or a SSBO.
+pub trait UniformBlock {        // TODO: `: Copy`, but unsized structs don't impl `Copy`
+    /// Checks whether the uniforms' layout matches the given block if `Self` starts at
+    /// the given offset.
+    fn matches(&BlockLayout, base_offset: usize) -> Result<(), LayoutMismatchError>;
+
+    /// Builds the `BlockLayout` corresponding to the current object.
+    fn build_layout(base_offset: usize) -> BlockLayout;
+}
+
+impl<T> UniformBlock for [T] where T: UniformBlock {
+    fn matches(layout: &BlockLayout, base_offset: usize)
+               -> Result<(), LayoutMismatchError>
+    {
+        if let &BlockLayout::DynamicSizedArray { ref content } = layout {
+            <T as UniformBlock>::matches(content, base_offset)
+                .map_err(|err| {
+                    LayoutMismatchError::MemberMismatch {
+                        member: "<dynamic array content>".to_string(),
+                        err: Box::new(err),
+                    }
+                })
+
+        } else if let &BlockLayout::Array { ref content, .. } = layout {
+            <T as UniformBlock>::matches(content, base_offset)
+                .map_err(|err| {
+                    LayoutMismatchError::MemberMismatch {
+                        member: "<dynamic array content>".to_string(),
+                        err: Box::new(err),
+                    }
+                })
+
+        } else {
+            Err(LayoutMismatchError::LayoutMismatch {
+                expected: layout.clone(),
+                obtained: <Self as UniformBlock>::build_layout(base_offset),
+            })
+        }
+    }
+
+    fn build_layout(base_offset: usize) -> BlockLayout {
+        BlockLayout::DynamicSizedArray {
+            content: Box::new(<T as UniformBlock>::build_layout(base_offset)),
+        }
+    }
 }